backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyyaml==6.0.1
pydantic==2.5.3